automation/requirements.txt
# automation/requirements.txt

google-generativeai
Pillow
